feat(memory): v0.18 plugin migration - restructure episodic memory as standalone plugin
- Migrate 12 source files from memory/ to ~/.hermes/plugins/episodic/
- Rename modules: episodic_store→store, episodic_provider→provider,
  session_journal→journal, plus config, compress, context_injector,
  extraction, merge, session_writer, temporal, wiki_distiller
- Add plugin.yaml with 11 tools and 5 hooks
- Create __init__.py with MemoryProvider registration
- Add memory → plugins/episodic symlink for backward compat
- Remove tool_calls parameter from sync_turn (zero core patches)
- Update test imports to new module paths (120 tests passing)
- Delete old plugins/memory/episodic/ bridge (superseded by standalone plugin)

Refs: v0.17→v0.18 structural refactor
