+++ memory
@@ -1,0 +1,1 @@
+/Users/telos/.hermes/plugins/episodic
