--- memory/__init__.py
@@ -1,1 +1,0 @@
-"""Sidecar Episodic Memory — persistent, searchable, lossless memory for Hermes."""
